Initial work on "variant generation" implementation.
The basic idea is that at the last minute before drawing, the code already has a lookup step where it computes the PSO to use, and that lookup step takes the `ProgramVars` as input!
So we can use that as an opportunity to specialize the chosen `Program` to the set of components on the `ProgramVars`.

Doing it this way is a bit gross (but not any grosser than the existing code, which was already using that opportunity to slam a few more `#define`s onto the program).
A better way would be to have an explicit cache that you look up based on the `Program` (entry point) and bound component classes (in the `ProgramVars`), but doing that would start to complicate the interaction with non-Spire shaders...

This hasn't been tested with multiple variants yet.

+++ Framework/Source/Graphics/Program.h
@@ -97,7 +97,7 @@ namespace Falcor
         bool isSpire() const { return mIsSpire; }
         ProgramReflection::SharedConstPtr getSpireReflector() const;
 
-        void setComponent(int index, SpireModule* componentClass);
+        void setComponent(size_t index, SpireModule* componentClass);
 
     protected:
         static const uint32_t kShaderCount = (uint32_t)ShaderType::Count;
